Moved DefaultDownloadStartupConfig to core
