+++ Tribler/Core/DownloadConfig.py
@@ -12,6 +12,8 @@
 # 3. Document your changes in API.py
 #
 #
+import copy
+import logging
 
 import os
 from types import StringType
@@ -207,6 +209,46 @@ class DownloadStartupConfig(DownloadConfigInterface):
 
     def copy(self):
         return DownloadStartupConfig(self.dlconfig.copy())
+
+
+class DefaultDownloadStartupConfig(DownloadStartupConfig):
+    """
+    This class is used to quickly get information about the default download startup config.
+    This is used for instance when adding a new torrent to the downloads. In this case, the default
+    download settings should be accessed and displayed to the user.
+    """
+    __single = None
+
+    def __init__(self, dlconfig=None):
+
+        if DefaultDownloadStartupConfig.__single:
+            raise RuntimeError("DefaultDownloadStartupConfig is singleton")
+        DefaultDownloadStartupConfig.__single = self
+
+        DownloadStartupConfig.__init__(self, dlconfig=dlconfig)
+
+        self._logger = logging.getLogger(self.__class__.__name__)
+
+    @staticmethod
+    def getInstance(*args, **kw):
+        if DefaultDownloadStartupConfig.__single is None:
+            DefaultDownloadStartupConfig(*args, **kw)
+        return DefaultDownloadStartupConfig.__single
+
+    @staticmethod
+    def delInstance(*args, **kw):
+        DefaultDownloadStartupConfig.__single = None
+
+    @staticmethod
+    def load(filename):
+        dlconfig = CallbackConfigParser()
+        dlconfig.read_file(filename)
+        return DefaultDownloadStartupConfig(dlconfig)
+
+    def copy(self):
+        config = CallbackConfigParser()
+        config._sections = {'downloadconfig': copy.deepcopy(self.dlconfig._sections['downloadconfig'])}
+        return DownloadStartupConfig(config)
 
 
 def get_default_dest_dir():
